requirements.txt
yfinance==0.2.48
pandas==2.2.3
numpy==2.1.2
matplotlib==3.9.2
python-dotenv==1.0.1
requests==2.32.3
pyyaml==6.0.2

